Feature: Moved Create to it's own view
