+++ coredatademo/ContentView.swift
@@ -10,6 +10,7 @@ import SwiftUI
 
 
 struct ContentView: View {
+    @State var showingChildView = false
     @FetchRequest(
         entity: Person.entity(),
         sortDescriptors: [
@@ -17,50 +18,26 @@ struct ContentView: View {
         ]
     ) var people: FetchedResults<Person>
     @Environment(\.managedObjectContext) var managedObjectContext
-    @State private var name: String = ""
-    @State private var age: Double = 10
-    var genders = ["Unspecified", "Male", "Female", "NonBinary"]
-    @State private var selectedGender = 0
+   
     var body: some View {
-        VStack{
+        NavigationView{
+            
             VStack{
-            TextField("Enter your name", text: $name)
-            Slider(value: $age, in: 10...90, step: 1.0)
-            Text(String(age))
-            Picker(selection: $selectedGender, label: Text("Please choose a gender")) {
-               ForEach(0 ..< genders.count) {
-                  Text(self.genders[$0])
-               }
-            }
-            Button(action: {
-                let person = Person(context: self.managedObjectContext)
-                person.name = self.name
-                person.age = Int64(Int(self.age))
-                person.gender = self.genders[self.selectedGender]
-                do {
-                    try self.managedObjectContext.save()
-                } catch {
-                    // handle the Core Data error
-                }
-                self.hideKeyboard()
-            }) {
-                Text("Insert Person")
-            }
-                Divider()
-            }
-            VStack{
-                Text("People").font(Font.headline)
-            NavigationView{
-                List{
-                    ForEach(people, id: \.self){ person in
-                Text(person.name ?? "Unknown")
-                    }.onDelete(perform: delete)
-                    
-                }.navigationBarItems(trailing: EditButton())
-            }
+                NavigationLink(destination: AddPeopleView(),
+                               isActive: self.$showingChildView)
+                { EmptyView() }
+                    .frame(width: 0, height: 0)
+                    .disabled(true)
+            Text("People").font(Font.headline)
+            List{
+            ForEach(people, id: \.self){ person in
+                NavigationLink(destination: Text(person.name ?? "Unknown")){
+                        Text(person.name ?? "Unknown")
+                    }
+                }.onDelete(perform: delete)
+            }.navigationBarItems(leading:Button(action:{ self.showingChildView = true }) { Text("Add") })
             }
         }
-        
     }
     func delete(at offsets: IndexSet) {
         for index in offsets {
